requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
unwrap==0.1.1
python-dotenv==0.19.0
psycopg2-binary>=2.8
flask-socketio==5.3.6
eventlet==0.36.1
jsonify==0.5
Flask-Dance[google]==7.1.0
oauthlib==3.2.2
